Flushing category additions for each record
Previously, if we were adding multiple categories with the same name in
a single batch, we'd get an error because the category names aren't
unique.

This is because no changes were committed to the database until the very
end, and therefore our duplicate detection in get_or_create would fail.
Flushing the session after adding each record ensures that duplicate
detection within the same session works as expected.