Added geocoding invalidation
While doing this, I also noticed that the system would error for
multiple scraper runs, which was fixed with a check to see if the
category to be added was already associated with the resource.

I also noticed an edge case with the joining of manual address fields -
if a field was empty, it would be included, which was alleviated with am
extra check against ''.  In addition, when performing such joins, each
field is stripped before it is included in the concatenation.