better CSV import modularization
The data_importer now has an enhanced role - it performs all the
filtering and normalization necessary to get correct RadRecords.  With
that in mind, the data_importer is now responsible for generating
RadRecords from the provided CSV file.

I also removed a bunch of unused stuff from the data_importer - I think
a lot of it was originally there to create categories, which is now
automatically handled by db_fun.

As part of this, I also generalized how we're filtering out invalid
column names.  Instead of just looking for an 'id' column and removing
it if it's there, we now grab a list of the fields available for the
RadRecord and filter out anything that's not in that list.