Tweaking OpenGraph/Twitter meta tags
The title used for OpenGraph and Twitter now uses the overall page title
and the URL will now be unique for each page.

The description is now customizable using an og_desc block and shared
between OpenGraph and Twitter. I'll start adding specific descriptions
to the appropriate pages.

The @radremedy Twitter page is now linked to appropriately. The Twitter
image was removed because it is not intended to be used for site-wide
logos (see https://dev.twitter.com/cards/types/summary). Similarly, the
Twitter domain is no longer supported.