Improved ratings for providers reviewed by user
The MVC action now passes in the date of the user's most
recently-visible review (if any). This is compared against the
last-reviewed date on the overall aggregates to determine if the current
user's review is incorporated in the results.

We now only show aggregate per-population ratings if there's more than
one person included or the user's review isn't included (either it
hasn't been calculated yet or there is no review by the user).