Adding feedback to quick location search
This required extensive tweaking of the styles - the main RAD logo was
overflowing its div, so I had to scale that back down. That had a whole
bunch of cascading effects in terms of form padding and icon sizing.

I also had to ensure that the main portion of the logo is
vertically-centered in the purple portion - it had extra whitespace at
the top.